Instruction for reading CPU time stamp counter.
The Time Stamp Counter (TSC) is a 64-bit register present on all x86 processors
since the Pentium. It counts the number of cycles since reset. The instruction
RDTSC returns the TSC in EDX:EAX. In x86-64 mode, RDTSC also clears the higher
32 bits of RAX and RDX. Its opcode is 0F 31.

We use the counter to measure the number of cycles of an if(1) statement.

Signed-off-by: Ilan Smith <lunnys@gmail.com>